atoi

+++ pointers_arrays_strings/100-atoi.c
@@ -11,7 +11,7 @@ int _atoi(char *s)
 {
 	int i = 0, sign = 1;
 	
-	long r = 0;
+	unsigned int r = 0;
 	
 	for (; s[i] != '\0'; ++i)
 	{
